buildImageArray creating an Image array from coloured coordinates

buildImageArray creating an Image array from coloured coordinates

buildImageArray(
  coordinates,
  sliceID = 1,
  invert = FALSE,
  na.rm = TRUE,
  resolution = 100,
  filterThreshold = 0.999,
  interpolation_type = 1,
  cores = 1,
  verbose = TRUE
)

Arguments

coordinates
a Vesalius data.frame with "barcodes","x","y","cc", "value","slice". Output from rgbPCA.

sliceID
integer - PC slice used in image building. Only one value!

invert
logical describing if colour pattern should be inverted (i.e 1-R,1-G,1-B)

na.rm
logical indicating if NA should be removed from bead list.

resolution
numeric (range 0 - 100) describing a percentage of original image size. Used to reduce image size.

filterThreshold
numeric (range 0 -1) describing the quantile threshold at which barcodes and tiles should be retained (seed details)

interpolation_type
Method of interpolation during image resizing:

-1

no interpolation: raw memory resizing.

0

no interpolation: additional space is filled according to boundary_conditions.

1

Nearest-neighbor interpolation

2

moving average interpolation

3

linear interpolation

4

grid interpolation

5

cubic interpolation

6

lanczos interpolation

cores
integer describing number of cores used (Default = 1)

verbose
logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value",and "tile".

Details

Vesalius converts PCA loading values into an RGB colour code associated with a coordinates on a Spatial Transcriptomic Assay. The buildImageArray converts coloured coordinates into an actual image via the use of Voronoi diagrams and Tile rasterisation.

Each barcodes coordinate is used as the center point of a voronoi tile and an artificial "box" is created surrounding barcode coordinates. This creates superfluous tiles that can be detrimental to further analysis. To remove excessive tiles, Vesalius takes a two step process. First, Vesalius removes any barcode that is too far away from other barcodes (likely stray barcodes). Second, Vesalius filters out tiles that are tied to the outer box and that exceed a certain area threshold. These triangles are nearly always related to boundary tiles. Both of these filtering steps are controlled via the filterThreshold argument.

A filterThreshold of 0.99 means that 99 % or barcodes and tile triangles will be retained.

Examples

if (FALSE) {
data(vesalius)
}