compareCells compute differential gene expression between selected territories for single cell type.

compareCells compute differential gene expression between selected territories for single cell type.

compareCells(
  image,
  counts,
  cells,
  seed = NULL,
  query = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  morphologyFactorSeed = 0,
  morphologyFactorQuery = 0,
  verbose = TRUE
)

Arguments

image

a dataframe containing territory information

counts

seurat object containing counts. Alternatively, matrix or sparse matrix. Colnames as barcodes and rownames as genes.

cells

character vector containing barcodes associated with cell type of interest.

seed

Integer or vector of integers describing territories to be included in group 1 for differential gene expression analysis.

query

Integer or vector of integers describing territories to be included in group 2 for differential gene expression analysis. Default = NULL

method

character describing the statistical test to use in order to extract differantial gene expression (currently only wilcox and t.test)

logFC

numeric describing minimum log fold change value for differential gene expression. Default set at 0.25.

pval

numeric for pval threshold. Default set at 0.05

minPct

numeric defining the minimum percentage of cells that should contain any given gene.

minBar

integer defining minimum number of barcodes in a territory.

morphologyFactorSeed

Integer or vector of integers describing growth or shrink rate of seed territories

morphologyFactorQuery

Integer or vector of integers describing growth or shrink rate of query territories

verbose

logical - progress message output

Value

A data.frame (tibble) containing differential gene expression as well p.value, logFC, seedPct (percentage of cells containing gene in first group), queryPct (percentage of cells containing gene in second group), seedTerritory (territory used as group 1) queryTerritory (territory used as group 2)

Details

compareCells provides a way to compare the expression of cells between different territories. If you have a cell type annotations associated to each of the barcodes present in your data, you can select your cell type of interest and parse the barcodes to the compareCell function.

Vesalius will internally subset your data and only compare barcodes present in your territories of interest for differential gene expressioon.

The cells argument should be supplied as a vector of barcodes containing all barcodes of the cell type of interest. Theoretically, you can supply more than one cell type in this vector but Vesalius will treat them as single cell type.

Examples