compareClusters computes differential gene expression between seurat clusters

compareClusters computes differential gene expression between seurat clusters

compareClusters(
  counts,
  seedCluster,
  queryCluster,
  seed = NULL,
  query = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  verbose = TRUE
)

Arguments

counts
seurat object containing ALL counts. Alternatively, matrix or sparse matrix of all counts. Colnames as barcodes and rownames as genes.

seedCluster
Seurat Object of seed clusters

queryCluster
Seurat Object of query clusters

seed
integer describing the Seurat cluster (in seed - group 1) to be used for Differential expression analysis. Default NULL

query
integer describing the Seurat cluster (in query - group 2) to be used for Differential expression analysis. Default NULL

method
character describing the statistical test to use in order to extract differential gene expression (currently only wilcox and t.test)

logFC
numeric describing minimum log fold change value for differential gene expression. Default set at 0.25.

pval
numeric for pval threshold. Default set at 0.05

minPct
numeric defining the minimum percentage of cells that should contain any given gene.

minBar
integer defining minimum number of barcodes in a territory.

verbose
logical - progress message output.

Value

A data.frame (tibble) containing differential gene expression as well p.value, logFC, seedPct (percentage of cells containing gene in first group), queryPct (percentage of cells containing gene in second group), seedTerritory (territory used as group 1) queryTerritory (territory used as group 2)

Details

Territory isolation provides a convenient way to investigate the finer detail of anatomy and cellular spatial distribution. In some case, two territories separated in 2D space may contain barcodes that cluster together and are described by similar cell type labels. The compareClusters functions compares clusters in different territories between each other. The count values are the overall count values that have been normalized, scaled and centered all together. This ensure that the barcodes are comparable. The territory isolation and clustering serves as a way to isolate a specific set of barcodes that will be used for differential gene expression analysis.

Examples

if (FALSE) {
data("Vesalius")

}