compareLayers computes differential gene expression between territory layers

compareLayers computes differential gene expression between territory layers

compareLayers(
  layers,
  counts,
  l1 = NULL,
  l2 = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  verbose = TRUE
)

Arguments

layers
data.frame containing layered territory information (See layerTerritory.edge and layerTerritory.concave)

counts
seurat object containing counts. Alternatively, matrix or sparse matrix. Colnames as barcodes and rownames as genes.

l1
integer or vector of integers indicating layers to be contained in group 1. Default NULL - If NULL will compare all layers independently

l2
integer or vector of integers indicating layers to be contained in group 2. If NULL will compare all layers independently

method
character describing the statistical test to use in order to extract differential gene expression (currently only wilcox and t.test)

logFC
numeric describing minimum log fold change value for differential gene expression. Default set at 0.25.

pval
numeric for pval threshold. Default set at 0.05

minPct
numeric defining the minimum percentage of cells that should contain any given gene.

minBar
integer defining minimum number of barcodes in a territory.

verbose
logical - progress message output.

Value

A data.frame (tibble) containing differential gene expression as well p.value, logFC, seedPct (percentage of cells containing gene in first group), queryPct (percentage of cells containing gene in second group), seedTerritory (territory used as group 1) queryTerritory (territory used as group 2)

Details

Territory isolation provides a convenient way to investigate the finer detail of anatomy and cellular spatial distribution. Image representation of territories provides a convenient way to manipulate territories such as dividing a territory into layers. Vesalius provides two method to layer a territory : layerTerritory.edge and layerTerritory.concave. The output of both function returns a data frame with layer information. The compareLayers compares a layer or layer sets between each other. If both l1 and l2 layers are left to NULL, each layer will be compared to each other layer on at a time. To compare one layer to all other layers, specific layer ID's should be specified.

Examples

if (FALSE) {
data("Vesalius")

}