equalizeHistogram image enhancement via colour histogram equalization.

equalizeHistogram image enhancement via colour histogram equalization.

equalizeHistogram(
  image,
  type = "BalanceSimplest",
  N = 1,
  smax = 1,
  sleft = 1,
  sright = 1,
  lambda = 0.1,
  up = 100,
  down = 10,
  invert = FALSE,
  verbose = TRUE
)

Arguments

image
data.frame - Vesalius formatted data.frame (i.e. barcodes, x,y,cc,value,...)

type
character - histogram EQ type. Select from: BalanceSimplest, EqualizePiecewise, SPE, EqualizeDP, EqualizeADP, ECDF (see details)

N
numeric describing how each colour channel will be mapped back to the image (Higher N = Higher greyscale contrast). Used with EqualizePiecewise

smax
numeric - upper limit if contrast stretching. Used with EqualizePiecewise

sleft
numeric - Range 0 - 100. Percentage of pixel to be saturated on the left side of the histogram. Used with BalanceSimplest

sright
numeric - Range 0 - 100. Percentage of pixel to be saturated on the right side of the histogram. Used with BalanceSimplest

lambda
numeric - strength of background correction. Used with SPE (Screened Poisson Equation).

up
numeric - color value threshold in the upper limit. Used with EqualizeDP.

down
numeric color value threshold in the lower limit. Used with EqualizeDP.

invert
logical - If TRUE, colours will be inverted i.e. 1 - colorValue (background set to 1 instead of 0).

verbose
logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value","tile",...

Details

Histogram equalization ensures that image details are amplified. In turn, territories may be extract with greater precision. We recommend balancing the histogram prior to smoothing.

For further details on each method described here, please refer to https://cran.r-project.org/web/packages/imagerExtra/vignettes/gettingstarted.html

Examples

if (FALSE) {
data(Vesalius)
}