exportRGB.csv export RGB coordinates as csv.

exportRGB.csv export RGB coordinates as csv.

exportRGB.csv(coordinates, file = NULL, slice = NULL, split = TRUE)

Arguments

coordinates
Vesalius data.frame with barcodes, x, y, cc, value,and slice.

file
character - file name or vector of files names if multiple slices will be exported.

slice
integer - slice to be exported

split
logical - If TRUE, slices will be exported in separately

Value

csv file at desired location containing rgb coloured coordinates.

Details

Export rgb colour code using slice information. If no file name or file names are provided, Vesalius will use its internal file name (pca_to_rgb.csv).

If one or more slice is provided, slice number will be included in file name. If split is true, indvidual files for each selected files will be created.