extractClusterMarkers compute differential gene expression between clusters and remaning barcodes

extractClusterMarkers compute differential gene expression between clusters and remaning barcodes

extractClusterMarkers(
  cluster,
  counts,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  verbose = TRUE
)

Arguments

cluster
a Seurat object containing clusters of an isolated territory

counts
seurat object containing counts. Alternatively, matrix or sparse matrix. Colnames as barcodes and rownames as genes.

method
character describing the statistical test to use in order to extract differential gene expression (currently only wilcox and t.test)

logFC
numeric describing minimum log fold change value for differential gene expression. Default set at 0.25.

pval
numeric for pval threshold. Default set at 0.05

minPct
numeric defining the minimum percentage of cells that should contain any given gene.

minBar
integer defining minimum number of barcodes in a territory.

verbose
logical - progress message output

Value

A data.frame (tibble) containing differential gene expression as well p.value, logFC, seedPct (percentage of cells containing gene in first group), queryPct (percentage of cells containing gene in second group), seedTerritory (territory used as group 1) queryTerritory (territory used as group 2)

Details

extractClusterMarkers compares clusters to all remaining barcodes. If a territory is isolated (see extractTerritories) for further analysis, Seurat can provide a cluster analysis of the isolated territory. Seurat will compared clusters between each other within the isolated territory. However, in some cases, it can be useful to compared the barcodes present in a Seurat cluster to all remaining barcodes. This provides overall differences in expression within the cluster as opposed to differential expression between isolated clusters.

Examples

if (FALSE) {
data("Vesalius")

}