extractMarkers compute differential gene expression between selected territories.

extractMarkers compute differential gene expression between selected territories.

extractMarkers(
  image,
  counts,
  seed = NULL,
  query = NULL,
  method = "wilcox",
  logFC = 0.25,
  pval = 0.05,
  minPct = 0.05,
  minBar = 10,
  morphologyFactorSeed = 0,
  morphologyFactorQuery = 0,
  verbose = TRUE
)

Arguments

image

a dataframe containing territory information

counts

seurat object containing counts. Alternatively, matrix or sparse matrix. Colnames as barcodes and rownames as genes.

seed

Integer or vector of integers describing territories to be included in group 1 for differential gene expression analysis.

query

Integer or vector of integers describing territories to be included in group 2 for differential gene expression analysis. Default = NULL

method

character describing the statistical test to use in order to extract differantial gene expression (currently only wilcox and t.test)

logFC

numeric describing minimum log fold change value for differential gene expression. Default set at 0.25.

pval

numeric for pval threshold. Default set at 0.05

minPct

numeric defining the minimum percentage of cells that should contain any given gene.

minBar

integer defining minimum number of barcodes in a territory.

morphologyFactorSeed

Integer or vector of integers describing growth or shrink rate of seed territories

morphologyFactorQuery

Integer or vector of integers describing growth or shrink rate of query territories

verbose

logical - progress message output

Value

A data.frame (tibble) containing differential gene expression as well p.value, logFC, seedPct (percentage of cells containing gene in first group), queryPct (percentage of cells containing gene in second group), seedTerritory (territory used as group 1) queryTerritory (territory used as group 2)

Details

extractMarkers compares a set of selected territories to another set of selected territories. If one of territory sets does not contain enough barcodes, extractMarkers will return NULL.

A seed territory (or territories) should always be provided. Query territories can be left as the default NULL. In this case, the seed territory will be compared to all remaining barcodes that are NOT present in the seed. Otherwise, seed territories will be compared to query territories after any morphological operations (see territoryMorphology.

extractMarkers provides a way to manipulate each set of territories independtly as described in morphologyFactorSeed and morphologyFactorQuery.

Examples

if (FALSE) {
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
markers <- extractMarkers(image, vesalius, seed = 1, query = 2)
}