layerTerritory.concave layers selected territories using concave hulling

layerTerritory.concave layers selected territories using concave hulling

layerTerritory.concave(
  image,
  seedTerritory = NULL,
  layerDepth = NULL,
  concavity = 1,
  length_threshold = 0,
  morphologyFactor = 3,
  captureRadius = 0.2,
  minBar = 10,
  verbose = TRUE
)

Arguments

image
data.frame - Vesalius formatted data.frame (i.e. barcodes, x,y,cc,value,cluster, territory)

seedTerritory
integer describing which territory should be layered

layerDepth
integer describing maximum number of layers

concavity
numeric describing concavity of the territory. 1 for highly detailed structures. At higher values the territory will be considered as a convex hull.

length_threshold
numeric describing the minimum distance between points to be considered during hulling. High values = simpler shapes

morphologyFactor
integer or vector of integers describing growth and/or shrink extent.

captureRadius
numeric - proportion of maximum distance between barcodes that will be used to pool barcodes together (range 0 - 1).

minBar
integer - minimum number of barcodes allowed in each territory

verbose
logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value","tile","cluster","territory" and "layer". Layer describes the layer to which a barcode belongs.

Details

To divide a territory into layers, one approach is to use concave hulling. This approach considers all barcodes locations (only center pixels) as points on a 2D grid and extracts the "outer" layer of barcodes.

This outer layer is extracted and defined as the first layer of the territory. The process is applied until no more barcodes can be pooled into a layer.

It should be noted that due to the geometrical nature of this approach, sub-territories are isolated prior to concave hulling to ensure that the hulling is only applied to contiguous barcodes (defined by captureRadius).

Examples

if (FALSE) {
data(Vesalius)
}