Embed 3D UMAP projections into RGB colour space

Embed 3D UMAP projections into RGB colour space

rgbUMAP(SO, pcs = 30, adjusted = FALSE, conserveSparse = TRUE, verbose = TRUE)

Arguments

SO

Seurat Object after normalisation, scaling and finding variable features.

pcs

number of Principle components to use for UMAP projections

adjusted

logical indicating if RGB code should be adjusted to colour contribution per channel.

conserveSparse

logical indicating if sparse matrix format should be conserved.

verbose

logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value","slice"

Details

Emedding 3D UMAP projections into the RGB color space is quite straight forward: Each UMAP dimension is extracted and min/max normalised to ensure that each value reanges between 0 and 1. The first dimension is assigned to red color channel, the second dimension to the green color channel and the third dimension assgined to the blue color channel. UMAP projections consider more varience per channel than PCA embeddings but this additional information in further compressed. Using PCA slice provides a way to visualize and explore different anatomical structure in more depth. Choosing between rgbPCA and rgbUMAP will depend on the question at hand and the depth of the analysis.

Examples

if (FALSE) {
data(Vesalius)
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbUMAP(image,pcs = 30)
}