territoryMorphing applies morphological operators to a set of territoriees

territoryMorphing applies morphological operators to a set of territoriees

territoryMorphing(territories, morphologyFactor, image, verbose = TRUE)

Arguments

territories

data.frame - Vesalius formatted data.frame (i.e. barcodes, x,y,cc,value,cluster, territory) only containing territories of interest.

morphologyFactor

integer or vector of integers describing growth and/or shrink extent.

image

data.frame - Vesalius formatted data.frame (i.e. barcodes, x,y,cc,value,cluster, territory) containing all territories.

verbose

logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value","tile","cluster", and "territory". This data.frame will contain the new territory after morphological operators and will contain barcodes associated with other territories and colour segments.

Details

Territory morphing can manipulate territories by growing, shrinking, filling, and cleaning territories. Growing = Positive integers - Territory will be dilated by x number of pixels Shrinking = Negative integers - Territory will be contracted by x number of pixels Filling = grow followed by shrink. Cleaning = shrink followed by grow.

Examples

if (FALSE) {
data(vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
ter <- image %>% filter(territory == 1)
# grow
g <- territoryMorphing(ter,morphologyFactor = 5, image = image)
# shrink
s <- territoryMorphing(ter,morphologyFactor = -5, image = image)
# fill
f <- territoryMorphing(ter,morphologyFactor = c(5,-5), image = image)
# clean
c <- territoryMorphing(ter,morphologyFactor = c(-5,5), image = image)
}