viewLayerExpression - plot gene expression in layered Vesalius territories

viewLayerExpression - plot gene expression in layered Vesalius territories

viewLayerExpression(image, counts, genes = NULL, normalise = TRUE, cex = 10)

Arguments

image

a Vesalius data frame containing barcodes, x, y, cc, value, cluster, and territory.

counts

count matrix - either matrix, sparse matrix or seurat object This matrix should contain genes as rownames and cells/barcodes as colnames

genes

character - gene of interest (only on gene at a time)

normalise

logical - If TRUE, gene expression values will be min/max normalised.

cex

numeric - font size modulator

Value

a ggplot object (geom_raster)

Details

Visualization of gene expression a layered territory. After isolation or a territory, Vesalius can apply territory layering. The expression shown here describes the mean expression in each layer.

Gene expression is shown "as is". This means that if no transformation is applied to the data then normalized raw count will be used for each layer.

If normalization and scaling have been applied, normalized counts will be used for each layer

This also applies to any data transformation applied on the count matrix or the Seurat object.

We recommend using Seurat Objects.

NOTE : You might be promoted to use geom_tiles instead of geom_raster. However - this will increase file size to unreasonable heights.

Examples

if (FALSE) {
data("vesalius")
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
image <- buildImageArray(image, sliceID=1)
# One segmentation round
image <- iterativeSegmentation.array(image)
image <- isolateTerritories.array(image, minBar = 5)
layer <- layerTerritory.edge(image, seedTerritory = 1)
g <- viewLayerExpression(layer, vesalius, genes = "Cst3")
}