Embed PCA loadings into RGB colour space

Embed PCA loadings into RGB colour space

rgbPCA(
  SO,
  slices = 1,
  adjusted = FALSE,
  rgbWeight = FALSE,
  countWeight = FALSE,
  conserveSparse = TRUE,
  verbose = TRUE
)

Arguments

SO

Seurat Object after normalisation, scaling and finding variable features.

slices

number of PCA slices to consider (Integer/numeric). Slice 1 will embed PC 1 to PC3, slice 2 will embed PC4 to PC6, etc

adjusted

logical indicating if RGB code should be adjusted to colour contribution per channel.

rgbWeight

logical indicating if RGB channels should be weighted by the variance associated to the PC it used for embedding.

countWeight

logical describing if colour embedding should be weighted by count number.

conserveSparse

logical indicating if sparse matrix format should be conserved.

verbose

logical - progress message output.

Value

Returns a Vesalius data.frame with "barcodes","x","y","cc", "value","slice" Slice represents the PC slice represented.

Details

The core concept of Vesalius is to embed PCA loading values into the RGB colour space. This is achieved by:

$$ B_{(c,i)} = \sum_{i=1} \mid L_(c,i) \mid $$

with B_((i,c)), barcode i in color channel c and L, the loading values associated to barcode i in color channel c. Notice that we take the absolute value of PCA loadings. In this case, we are only interested in variance and not direction of variance.

Barcode values are then min/max normalised to ensure that all colour codes ranged from 0 to 1.

Examples

if (FALSE) {
data(Vesalius)
# Seurat pre-processing
image <- NormalizeData(vesalius)
image <- FindVariableFeatures(image, nfeatures = 2000)
image <- ScaleData(image)
# converting to rgb
image <- rgbPCA(image,slices = 1)
}